requirements.txt
fastapi
uvicorn
pillow
pytesseract
openai
python-multipart
pdf2image
requests
gTTS
